requirements.txt
pip~=22.1.2
wheel~=0.36.2
Django~=4.0.5
sqlparse~=0.4.2
asgiref~=3.5.2
setuptools~=57.4.0
djangorestframework~=3.13.1
django-filter~=21.1
psycopg2>=2.8